# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/tokens": {
            "get": {
                "description": "retrieve the list of supported tokens",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tokens"
                ],
                "summary": "Get the tokens list",
                "operationId": "get-tokens",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/tokens.Token"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "create a new token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tokens"
                ],
                "summary": "Create a new token",
                "operationId": "post-token",
                "parameters": [
                    {
                        "description": "Token object that needs to be added",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SaveTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    }
                }
            },
            "patch": {
                "description": "update a token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tokens"
                ],
                "summary": "Update a token",
                "operationId": "patch-token",
                "parameters": [
                    {
                        "description": "Token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.EditTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/tokens/{id}": {
            "get": {
                "description": "retrieve a supported token by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tokens"
                ],
                "summary": "Get a token",
                "operationId": "get-token-by-id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tokens.Token"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete a token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tokens"
                ],
                "summary": "Delete a token",
                "operationId": "delete-token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Result"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "tokens.Blockchain": {
            "type": "object",
            "properties": {
                "abbr": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "main_token": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "tokens.Token": {
            "type": "object",
            "properties": {
                "abbr": {
                    "type": "string"
                },
                "blockchain": {
                    "$ref": "#/definitions/tokens.Blockchain"
                },
                "blockchain_id": {
                    "type": "string"
                },
                "contract": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "precision": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "types.EditTokenRequest": {
            "type": "object",
            "properties": {
                "abbr": {
                    "type": "string",
                    "minLength": 2,
                    "example": "BTC"
                },
                "blockchain": {
                    "type": "string",
                    "example": "66f6fe7eccc6398d39e981f9"
                },
                "contract": {
                    "type": "string",
                    "example": "0x1234567890"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "TOKEN-NAME"
                },
                "precision": {
                    "type": "integer",
                    "example": 18
                },
                "type": {
                    "type": "string",
                    "example": "NATIVE"
                }
            }
        },
...
        "types.ErrorMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "types.Result": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                }
            }
        },
...
        "types.SaveTokenRequest": {
            "type": "object",
            "required": [
                "abbr",
                "blockchain",
                "contract",
                "precision",
                "type"
            ],
            "properties": {
                "abbr": {
                    "type": "string",
                    "minLength": 2,
                    "example": "BTC"
                },
                "blockchain": {
                    "type": "string",
                    "example": "66f6fe7eccc6398d39e981f9"
                },
                "contract": {
                    "type": "string",
                    "example": "0x1234567890"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "TOKEN-NAME"
                },
                "precision": {
                    "type": "integer",
                    "example": 18
                },
                "type": {
                    "type": "string",
                    "example": "NATIVE"
                }
            }
        },
...
    }
}